Add a task to merge the vcfs output by HaplotypeCaller
The final output is a single gvcf file combining the scattered output
vcf files from the HaplotypeCaller task, as well as the index for this
file.
Also removed reference to res_dir; moving files into a results directory
is not necessary in a cloud environment, and files are output directly.
If files must be placed in a specially named output directory another
simple task could be added that moves the files into this folder and
tars the result.